Separate Unity executable resolution from launch flow
Keep launch orchestration focused and satisfy the repository source-file size guard without relying on non-formatted whitespace.
