+++ cli/dispatcher/internal/dispatcher/launch.go
@@ -441,37 +441,6 @@ func unityExecutableCandidates(version string) []string {
 		return []string{}
 	}
 }
-func windowsUnityExecutableCandidates(version string) []string {
-	candidates := []string{}
-	for _, base := range []string{
-		os.Getenv("ProgramFiles"),
-		os.Getenv("ProgramFiles(x86)"),
-		os.Getenv("LOCALAPPDATA"),
-		`C:\Program Files`,
-	} {
-		if base == "" {
-			continue
-		}
-		candidates = append(candidates, filepath.Join(base, "Unity", "Hub", "Editor", version, "Editor", "Unity.exe"))
-	}
-	return candidates
-}
-func readUnityEditorVersion(projectRoot string) (string, error) {
-	content, err := os.ReadFile(filepath.Join(projectRoot, projectVersionFilePath))
-	if err != nil {
-		return "", err
-	}
-	matches := editorVersionPattern.FindStringSubmatch(string(content))
-	if len(matches) != 2 {
-		return "", fmt.Errorf("unity editor version not found in %s", projectVersionFilePath)
-	}
-	version := strings.TrimSpace(matches[1])
-	if version == "" {
-		return "", fmt.Errorf("unity editor version is empty in %s", projectVersionFilePath)
-	}
-	return version, nil
-}
-
 func killUnityProcess(pid int) error {
 	process, err := os.FindProcess(pid)
 	if err != nil {
